# core-lending-provisioning-risk-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-provisioning-risk
  description: Provisioning Risk Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/provisioning-cases/{provisioningCaseId}/risk-assessments/{riskAssessmentId}:
    get:
      tags:
        - RiskAssessment
      summary: Get a risk assessment by ID
      operationId: getById_2
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riskAssessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RiskAssessmentDTO'
    put:
      tags:
        - RiskAssessment
      summary: Update a risk assessment
      operationId: update_2
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riskAssessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAssessmentDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RiskAssessmentDTO'
    delete:
      tags:
        - RiskAssessment
      summary: Delete a risk assessment
      operationId: delete_2
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riskAssessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/provisioning-cases/{provisioningCaseId}/risk-assessments:
    get:
      tags:
        - RiskAssessment
      summary: List or search risk assessments for a provisioning case
      operationId: findAll_2
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.provisioningCaseId
          in: query
          description: Exact filter for provisioning case id
          required: false
          schema:
            type: string
        - name: filters.pdValue
          in: query
          description: Exact filter for pd value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[pdValue].from
          in: query
          description: Filter pd value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[pdValue].to
          in: query
          description: Filter pd value to value
          required: false
          schema:
            type: string
        - name: filters.lgdValue
          in: query
          description: Exact filter for lgd value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[lgdValue].from
          in: query
          description: Filter lgd value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[lgdValue].to
          in: query
          description: Filter lgd value to value
          required: false
          schema:
            type: string
        - name: filters.eadValue
          in: query
          description: Exact filter for ead value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eadValue].from
          in: query
          description: Filter ead value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eadValue].to
          in: query
          description: Filter ead value to value
          required: false
          schema:
            type: string
        - name: filters.modelVersion
          in: query
          description: Exact filter for model version
          required: false
          schema:
            type: string
        - name: filters.scenarioCode
          in: query
          description: Exact filter for scenario code
          required: false
          schema:
            type: string
            enum:
              - BASE
              - OPTIMISTIC
              - ADVERSE
              - SEVERE
        - name: filters.assessmentDate
          in: query
          description: Exact filter for assessment date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[assessmentDate].from
          in: query
          description: Filter assessment date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[assessmentDate].to
          in: query
          description: Filter assessment date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.details
          in: query
          description: Exact filter for details
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRiskAssessmentDTO'
    post:
      tags:
        - RiskAssessment
      summary: Create a new risk assessment
      operationId: create_2
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAssessmentDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RiskAssessmentDTO'
...
components:
  schemas:
...
    RiskAssessmentDTO:
      required:
        - assessmentDate
        - eadValue
        - lgdValue
        - modelVersion
        - pdValue
        - provisioningCaseId
        - scenarioCode
      type: object
      properties:
        riskAssessmentId:
          type: string
          format: uuid
          readOnly: true
        provisioningCaseId:
          type: string
          format: uuid
        pdValue:
          maximum: 1
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        lgdValue:
          maximum: 1
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        eadValue:
          type: number
        modelVersion:
          maxLength: 100
          minLength: 0
          type: string
        scenarioCode:
          type: string
          enum:
            - BASE
            - OPTIMISTIC
            - ADVERSE
            - SEVERE
        assessmentDate:
          type: string
          format: date-time
        details:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseRiskAssessmentDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/RiskAssessmentDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...